+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Add support for the geNomad database v1.8, which introduced an additional column in the metadata table.
 
 ## [1.9.0] - 2025-01-25
 ### Changed